+++ cipher/aes.go
@@ -1,26 +1,30 @@
-package command
+package cipher
 
 import (
 	cipherAES "github.com/mvisonneau/s5/cipher/aes"
 
 	log "github.com/sirupsen/logrus"
-	"github.com/urfave/cli"
 )
 
-type aes struct {
+type AES struct {
 	client *cipherAES.Client
 }
 
-func (a *aes) configure(ctx *cli.Context) (err error) {
-	a.client, err = cipherAES.Init(
+func NewAES(key string) (*AES, error) {
+	c, err := cipherAES.Init(
 		&cipherAES.Config{
-			Key: ctx.String("key"),
+			Key: key,
 		},
 	)
-	return
+
+	if err != nil {
+		return nil, err
+	}
+
+	return &AES{c}, nil
 }
 
-func (a *aes) cipher(value string) (string, error) {
+func (a *AES) Cipher(value string) (string, error) {
 	log.Debug("Ciphering using AES")
 	ciphered, err := a.client.Cipher(value)
 	if err != nil {
@@ -29,7 +33,7 @@ func (a *aes) cipher(value string) (string, error) {
 	return ciphered, nil
 }
 
-func (a *aes) decipher(value string) (string, error) {
+func (a *AES) Decipher(value string) (string, error) {
 	log.Debugf("Deciphering '%s' using AES", value)
 	plain, err := a.client.Decipher(value)
 	if err != nil {