+++ cipher/gcp.go
@@ -1,26 +1,30 @@
-package command
+package cipher
 
 import (
 	cipherGCP "github.com/mvisonneau/s5/cipher/gcp"
 
 	log "github.com/sirupsen/logrus"
-	"github.com/urfave/cli"
 )
 
-type gcp struct {
+type GCP struct {
 	client *cipherGCP.Client
 }
 
-func (g *gcp) configure(ctx *cli.Context) (err error) {
-	g.client, err = cipherGCP.Init(
+func NewGCP(kmsKeyName string) (*GCP, error) {
+	c, err := cipherGCP.Init(
 		&cipherGCP.Config{
-			KmsKeyName: ctx.String("kms-key-name"),
+			KmsKeyName: kmsKeyName,
 		},
 	)
-	return
+
+	if err != nil {
+		return nil, err
+	}
+
+	return &GCP{c}, nil
 }
 
-func (g *gcp) cipher(value string) (string, error) {
+func (g *GCP) Cipher(value string) (string, error) {
 	log.Debug("Ciphering using a GCP KMS key")
 	ciphered, err := g.client.Cipher(value)
 	if err != nil {
@@ -29,7 +33,7 @@ func (g *gcp) cipher(value string) (string, error) {
 	return ciphered, nil
 }
 
-func (g *gcp) decipher(value string) (string, error) {
+func (g *GCP) Decipher(value string) (string, error) {
 	log.Debugf("Deciphering '%s' using a GCP KMS key", value)
 	plain, err := g.client.Decipher(value)
 	if err != nil {