+++ cipher/vault.go
@@ -1,26 +1,30 @@
-package command
+package cipher
 
 import (
 	cipherVault "github.com/mvisonneau/s5/cipher/vault"
 
 	log "github.com/sirupsen/logrus"
-	"github.com/urfave/cli"
 )
 
-type vault struct {
+type Vault struct {
 	client *cipherVault.Client
 }
 
-func (v *vault) configure(ctx *cli.Context) (err error) {
-	v.client, err = cipherVault.Init(
+func NewVault(transitKey string) (*Vault, error) {
+	c, err := cipherVault.Init(
 		&cipherVault.Config{
-			Key: ctx.String("transit-key"),
+			Key: transitKey,
 		},
 	)
-	return
+
+	if err != nil {
+		return nil, err
+	}
+
+	return &Vault{c}, nil
 }
 
-func (v *vault) cipher(value string) (string, error) {
+func (v *Vault) cipher(value string) (string, error) {
 	log.Debug("Ciphering using Vault transit key")
 	ciphered, err := v.client.Cipher(value)
 	if err != nil {
@@ -29,7 +33,7 @@ func (v *vault) cipher(value string) (string, error) {
 	return ciphered, nil
 }
 
-func (v *vault) decipher(value string) (string, error) {
+func (v *Vault) Decipher(value string) (string, error) {
 	log.Debugf("Deciphering '%s' using Vault transit key", value)
 	plain, err := v.client.Decipher(value)
 	if err != nil {